fix: split msg server files
