+++ x/house/types/message_withdraw.go
@@ -8,7 +8,7 @@ import (
 
 const typeMsgWithdraw = "withdraw"
 
-var _ sdk.Msg = &MsgDeposit{}
+var _ sdk.Msg = &MsgWithdraw{}
 
 // NewMsgWithdraw creates the new input for withdrawal of a deposit
 func NewMsgWithdraw(creator string, sportEventUID string, amount sdk.Int, participationIndex uint64, mode WithdrawalMode) *MsgWithdraw {
